fix(app): fix bug where savereminderform would incorrectly display with no content
Update app.tsx to no longer push directly to the SaveReminderForm if the user is on the /submissions
page. This push was breaking some state, causing the SaveReminderForm to appear empty even if a
reminder had previously been saved for that specific problem. Remove unneeded code from DomEvaluator
as part of code cleanup.